Updated openai project use Official golang sdk

+++ openai-web/models/completion.go
@@ -7,12 +7,12 @@ import (
 type CompletionModel struct {
 	Model            string   `json:"model"`
 	Prompt           string   `json:"prompt"`
-	MaxTokens        int32    `json:"max_tokens"`
-	Temperature      float32  `json:"temperature" valid:"ValidateTemperature"`
-	Top_P            float32  `json:"top_p" valid:"ValidateTemperature"`
-	FrequencyPenalty float32  `json:"frequency_penalty" valid:"ValidatePenalty"`
-	PresencePenalty  float32  `json:"presence_penalty" valid:"ValidatePenalty"`
-	BestOf           int32    `json:"best_of" valid:"Range(1, 20)"`
+	MaxTokens        int64    `json:"max_tokens"`
+	Temperature      float64  `json:"temperature" valid:"ValidateTemperature"`
+	Top_P            float64  `json:"top_p" valid:"ValidateTemperature"`
+	FrequencyPenalty float64  `json:"frequency_penalty" valid:"ValidatePenalty"`
+	PresencePenalty  float64  `json:"presence_penalty" valid:"ValidatePenalty"`
+	BestOf           int64    `json:"best_of" valid:"Range(1, 20)"`
 	Echo             bool     `json:"echo"`
 	Logprobs         int      `json:"logprobs"`
 	Stop             []string `json:"stop"`
@@ -26,7 +26,7 @@ func init() {
 }
 
 var validateTemperature validation.CustomFunc = func(v *validation.Validation, obj interface{}, key string) {
-	temperature, ok := obj.(float32)
+	temperature, ok := obj.(float64)
 
 	if !ok {
 		return
@@ -38,7 +38,7 @@ var validateTemperature validation.CustomFunc = func(v *validation.Validation, o
 }
 
 var validatePenalty validation.CustomFunc = func(v *validation.Validation, obj interface{}, key string) {
-	temperature, ok := obj.(float32)
+	temperature, ok := obj.(float64)
 
 	if !ok {
 		return
